studiobot/requirements.txt
aiogram==3.20.0
sqlalchemy==2.0.41
asyncpg==0.30.0
celery==5.5.3
redis==6.2.0
# easyocr==1.7.2
pytesseract==0.3.13
PyMuPDF==1.26.3
pdf2image==1.17.0
Pillow==11.3.0
gevent==25.5.1
